+++ main/models.py
@@ -21,7 +21,7 @@ class Goal(models.Model):
     created = models.DateField()    # 목표 글 작성 날짜
     start_date = models.DateField()     # 목표 시행 날짜
     fee = models.IntegerField(default=500)      # 참가비
-    value = models.IntegerField(null=True, blank=True)     # 인증 방식이 수치인 경우만
+    value = models.FloatField(null=True, blank=True)     # 인증 방식이 수치인 경우만
     unit = models.CharField(max_length=10, null=True, blank=True)     # 인증 방식이 수치인 경우만
     criteria = models.BooleanField(default=True)    # 인증 방식이 수치인 경우만. true면 이상, false면 이하로 처리
     member = models.ManyToManyField(User, related_name='members')    # 참여 인원
